address comments: remove ledger extension and move notaries back to membership module
