--- ledger/ledger-common/src/test/java/net/corda/v5/ledger/common/transaction/GroupParametersExtensionJavaApiTest.java
@@ -1,30 +1,0 @@
-package net.corda.v5.ledger.common.transaction;
-
-import net.corda.v5.ledger.common.GroupParametersExtension;
-import net.corda.v5.ledger.common.NotaryInfo;
-import net.corda.v5.membership.GroupParameters;
-import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.Test;
-import java.util.List;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-import static org.mockito.kotlin.MatchersKt.eq;
-
-public class GroupParametersExtensionJavaApiTest {
-
-    private final GroupParameters groupParameters = mock(GroupParameters.class);
-    private final NotaryInfo notaryInfo = mock(NotaryInfo.class);
-    private final List<NotaryInfo> notaryInfoList = List.of(notaryInfo);
-
-    @Test
-    public void getNotaryServiceParty() {
-        String notariesKey = "corda.notaries";
-        when(groupParameters.parseList(eq(notariesKey), eq(NotaryInfo.class))).thenReturn(notaryInfoList);
-
-        var result = GroupParametersExtension.getNotaries(groupParameters);
-
-        Assertions.assertThat(result).isNotNull();
-        Assertions.assertThat(result).isEqualTo(notaryInfoList);
-    }
-}
-
