+++ membership/src/main/kotlin/net/corda/v5/membership/NotaryInfo.kt
@@ -1,4 +1,4 @@
-package net.corda.v5.ledger.common
+package net.corda.v5.membership
 
 import net.corda.v5.base.annotations.CordaSerializable
 import net.corda.v5.base.types.MemberX500Name
